requirements.txt
# requirements.txt 
requests
python-dotenv
